Qualify direct recovery against release-profile renderer builds
