+++ crates/noon-web/Cargo.toml
@@ -10,6 +10,8 @@ crate-type = ["cdylib", "rlib"]
 
 [features]
 default = ["renderer"]
+# Explicit release-profile qualification only; production builds omit the fixtures.
+renderer-smoke = ["renderer"]
 renderer = [
     "dep:noon-render-wgpu",
     "dep:noon-text-render-wgpu",
