+++ scripts/build-web-demo.sh
@@ -127,6 +127,10 @@ if [[ "${NOON_WASM_SKIP_OPT:-0}" == "1" ]]; then
   wasm_pack_args+=(--no-opt)
 fi
 
+if [[ "${NOON_RENDERER_SMOKE:-0}" == "1" ]]; then
+  wasm_pack_args+=(--features renderer-smoke)
+fi
+
 worker_pid=""
 if (( parallel_worker == 1 )); then
   node scripts/build-python-worker.mjs &
